requirements.txt
flask>=2.0.1
pillow>=9.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
werkzeug>=2.0.0
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
ultralytics>=8.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
sentencepiece>=0.1.99
schedule>=1.1.0
torchvision>=0.15.0
